fortuna_android/fine-tuning/requirements.txt
# SmolVLM Fine-tuning Requirements
# Install with: pip install -r requirements.txt

# Core ML frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Dataset and processing
datasets>=2.15.0
Pillow>=10.0.0
requests>=2.31.0

# OpenAI API for labeling
openai>=1.6.0

# Training utilities
wandb>=0.16.0
trl>=0.7.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
